Move RunShowElectionRecord to jvm cli. readMixnetJsonBallots can use List<List<List<String>>> without needing an object.
