+++ egklib/src/jvmTest/kotlin/electionguard/cli/RunShowElectionRecordTest.kt
@@ -1,0 +1,41 @@
+package electionguard.cli
+
+import kotlin.test.Test
+
+class RunShowElectionRecordTest {
+
+    @Test
+    fun testShowElectionRecordTest() {
+        RunShowElectionRecord.main(
+            arrayOf(
+                "-in", "src/commonTest/data/mixnet/working/eg/keyceremony",
+                "-show", "manifest",
+            )
+        )
+    }
+
+    @Test
+    fun testShowElectionRecordDetailsTest() {
+        RunShowElectionRecord.main(
+            arrayOf(
+                "-in", "src/commonTest/data/mixnet/working/eg/keyceremony",
+                "-show", "manifest",
+                "--details",
+            )
+        )
+    }
+
+    @Test
+    fun testShowElectionRecordBallotStyleTest() {
+        RunShowElectionRecord.main(
+            arrayOf(
+                "-in", "src/commonTest/data/mixnet/working/eg/keyceremony",
+                "-show", "manifest",
+                "-ballotStyle", "ballot-style-1",
+                "--details",
+            )
+        )
+    }
+
+}
+
